Restricted access for CRUD
Added restrictions for authenticated users and admin only.

+++ src/Controller/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\RoleEnum;
 use App\Entity\User;
 use App\Form\UserForm;
 use App\Repository\UserRepository;
@@ -18,6 +19,9 @@ final class UserController extends AbstractController
     #[Route(name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
+        // Only authenticated users should access this
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED', null, "An unauthenticated user tried to access route app_user_index.");
+        
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -26,6 +30,9 @@ final class UserController extends AbstractController
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $passwordHasher): Response
     {
+        // Only authenticated users should access this
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED', null, "A user tried to access route app_user_new without authorization.");
+        
         $user = new User();
         $form = $this->createForm(UserForm::class, $user);
         $form->handleRequest($request);
@@ -51,6 +58,9 @@ final class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        // Only authenticated users should access this
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED', null, "A user tried to access route app_user_show without authorization.");
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -59,6 +69,9 @@ final class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        // Only authenticated users should access this
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED', null, "A user tried to access route app_user_edit without authorization.");
+
         $form = $this->createForm(UserForm::class, $user);
         $form->handleRequest($request);
 
@@ -77,6 +90,9 @@ final class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        // Only authenticated users should access this
+        $this->denyAccessUnlessGranted(RoleEnum::ADMIN->value, null, "A user tried to access route app_user_delete without authorization.");
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
